[snap/frontend]: move generateAccountsState method to testHelper, and minor issue
