+++ snap/frontend/components/AccountInfo/index.jsx
@@ -7,15 +7,15 @@ const AccountInfo = () => {
 	const { walletState } = useWalletContext();
 	const { accounts, selectedAccount } = walletState;
 
-	const [ isOpenAccountListModalBox, setOpenAccountListModalBox ] = useState(false);
+	const [ accountListModalBoxVisible, setAccountListModalBoxVisible ] = useState(false);
 
 	const handleAccountListModalBox = () => {
-		setOpenAccountListModalBox(!isOpenAccountListModalBox);
+		setAccountListModalBoxVisible(!accountListModalBoxVisible);
 	};
 
 	return (
 		<>
-			<AccountListModalBox accounts={accounts} isOpen={isOpenAccountListModalBox} onRequestClose={setOpenAccountListModalBox} />
+			<AccountListModalBox accounts={accounts} isOpen={accountListModalBoxVisible} onRequestClose={setAccountListModalBoxVisible} />
 
 			<div className='flex flex-col items-center justify-center p-2'>
 				<div role='profile-image' className="rounded-full w-16 h-16 bg-gray-300" onClick={handleAccountListModalBox}/>
